+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
-
+### Added
+- Added option for the agent to open the generated Word documents.
 
 
 ## [v0.2.0] - 2025-05-26