+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+
+## [v0.3.0] - 2025-05-28
 ### Added
 - Added option for the agent to open the generated Word documents.
 - Added ability to broadcast messages to the user.